Implement GET and SET commands
A map is used to store key-value pairs in a
concurrency safe manner.
There is a scope to improve the performance here.
It will probably be more performant to store entire
bulk strings in the map rather than only the data part.
When doing, the entire bulk string can returned back to the
client. Thereby, avoiding a need to serialise the data.